Cleanup InfoView Feature
